H2: Solve problem 5, reformat 4 floats
Write solution for problem 5. Problem 4 percentage calculations (1.076) reformated from:

(float)1.076 ---> 1.076(f)

To be in line with Java Data Types class handout.

Problem 5:

All years that are evenly divisible by 400 are always a leap year. They are also a leap year if  they are evenly divisible by 4 AND are not evenly divisible by 100. For example, 1600 was a  leap year because it is evenly divisible by 400. Likewise 1988 was a leap year because it was  evenly divisible by 4 AND not evenly divisible by 100. Write an algorithm to solve this problem  for entering any year, providing a positive integer.

+++ HW2/src/Problem_4.java
@@ -8,10 +8,10 @@ public class Problem_4 {
         float Base_Annual_Salary = Base_Annual_Salary_Input.nextFloat();
         Base_Annual_Salary_Input.close();
 
-        float Retroactive_Pay = ((Base_Annual_Salary / 2) * (float)1.076) - (Base_Annual_Salary / 2);
+        float Retroactive_Pay = ((Base_Annual_Salary / 2) * 1.076f) - (Base_Annual_Salary / 2);
         System.out.println("Retroactive Pay --> $" + Retroactive_Pay);
 
-        float Adjusted_Annual_Salary = Base_Annual_Salary * (float)1.076;
+        float Adjusted_Annual_Salary = Base_Annual_Salary * 1.076f;
         System.out.println("New Annual Salary --> $" + Adjusted_Annual_Salary);
 
         float Adjusted_Monthly_Salary = Adjusted_Annual_Salary / 12;
